requirements.txt
Django>=3.2,<4.0
psycopg2-binary>=2.8
djangorestframework>=3.12
django-cors-headers>=3.5
pillow
sqlparse>=0.4
